fix: internal/build silently dropped bundle-declared plugins (CI regression)
bundle.Load() was only wired into 'nself bundle's PersistentPreRunE. Every
OTHER command that reaches internal/build's manifest expansion (build,
start, restart, stop, ops, status) never populated the resolver at all, so
a manifest's 'bundles:' entries silently expanded to nothing regardless of
slug spelling -- not an alias-direction bug, a missing-Load() bug. Root
cause confirmed by reproducing the CI failure locally: the expansion
produced 0 plugins from the bundle, not 13-vs-14 from a wrong alias.

Fix: bundle.EnsureLoaded(ctx) -- idempotent, no-op once any caller in the
process has already loaded the registry -- called at the actual expansion
site in DeclaredPlugins() before iterating bundles. On failure it logs a
loud slog.Warn naming the bundles that could not be expanded, per this
file's own PCI charter (never silently drop a declared plugin).

Also: TestLoadProjectManifest_FlatPluginsAndBundleExpansion's expected
sentry-bundle plugin count (13) predated ADR-P6-03's real bundles.json
content (14, including nself-stripe per Ruling 2) -- updated 13->14/15,
and the test now seeds a local bundles.json fixture via bundle.LoadBytes
in a TestMain so internal/build's tests stay deterministic and offline.

Split registry.go (302 lines) into registry.go + registry_cache.go, and
manifest.go (310 lines) into manifest.go + manifest_resolve.go to stay
under the internal/repoqa 300-line/file cap after these additions.

Verified: go build/vet/test -mod=vendor (matching CI's GOFLAGS) all green
on darwin/arm64 native plus cross-compiled vet for linux/amd64,
linux/arm64, windows/amd64, darwin/amd64. golangci-lint -mod=vendor:
12 findings, identical set to a clean origin/main checkout -- zero
findings introduced by this diff.
